Refactor macho ctx kit
